refactor: 상수화 관련 리팩토링 - 여러 개의 상수화 파일들을 하나의 파일로 합침 - 하나의 파일에서 관리하는 것이 더 가독성 있다고 생각하여 진행하였습니다
