+++ src/constants/constants.js
@@ -1,6 +1,5 @@
 export const STRING = {
   PRODUCTS: 'products',
-  CHARGE_AMOUNT: 'chargeAmount',
   COINS: 'coins',
   INPUT_CHARGE_AMOUNT: 'inputChargeAmount',
   CURRENT_TAB: 'currentTab',
@@ -9,7 +8,8 @@ export const STRING = {
   TR: 'tr',
 };
 
-const NUMBER = {
+export const NUMBER = {
+  MINUS_ONE: -1,
   ZERO: 0,
   ONE: 1,
   FOUR: 4,
@@ -17,6 +17,7 @@ const NUMBER = {
   FIFTY: 50,
   HUNDRED: 100,
   FIVE_HUNDRED: 500,
+  THOUSAND: 1000,
 };
 
 export const COIN_UNIT = [NUMBER.FIVE_HUNDRED, NUMBER.HUNDRED, NUMBER.FIFTY, NUMBER.TEN];
@@ -32,4 +33,19 @@ export const TAB = {
   ADD: 'add',
   MANAGE: 'manage',
   PURCHASE: 'purchase',
+};
+
+export const ERROR_MESSAGE = {
+  EMPTY_PRODUCT_NAME: '상품명을 입력해주세요',
+  DUPLICATED_PRODUCT_NAME: '중복되지 않는 상품명을 기입해주세요',
+  EMPTY_PRODUCT_PRICE: '가격을 입력해주세요',
+  EMPTY_PRODUCT_QUANTITY: '수량을 입력해주세요',
+  EMPTY_CHARGE_AMOUNT: '금액을 입력해주세요',
+  NOT_OVER_HUNDRED: '100원이상의 가격을 입력해주세요',
+  NOT_OVER_TEN: '10이상의 숫자를 입력해주세요',
+  NOT_OVER_ONE: '1이상의 숫자를 입력해주세요',
+  NOT_DIVIDED_BY_TEN_PRICE: '10원으로 나누어 떨어지는 가격을 입력해주세요',
+  NOT_DIVIDED_BY_TEN: '10으로 나누어 떨어지는 숫자를 입력해주세요',
+  NOT_ENOUGH_CHARGE: '보유하신 금액으로 해당 상품을 구매하실 수 없습니다',
+  NOT_ENOUGH_QUANTITY: '해당 상품은 품절이므로 구매하실 수 없습니다',
 };
