--- src/constants/number.js
@@ -1,13 +1,0 @@
-const NUMBER = {
-  MINUS_ONE: -1,
-  ZERO: 0,
-  ONE: 1,
-  FOUR: 4,
-  TEN: 10,
-  FIFTY: 50,
-  HUNDRED: 100,
-  FIVE_HUNDRED: 500,
-  THOUSAND: 1000,
-};
-
-export default NUMBER;
